requirements.txt
matplotlib==3.7.1
numpy~=1.24.3
pandas~=2.0.2
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.3.2
traitlets==5.9.0
tzdata==2023.3
uri-template==1.2.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
widgetsnbextension==4.0.7
statsmodels==0.14.0
xgboost==1.7.5
imbalanced-learn==0.10.1
scikit-optimize==0.9.0
geopandas==0.13.0
shapely==2.0.1
greenlet==1.1.2
mlflow==2.3.2